requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
codecov==2.0.22
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
dj-database-url==0.5.0
Django==3.0.5
djangorestframework==3.11.0
drf-yasg==1.17.1
entrypoints==0.3
fakeredis==1.4.1
flake8==3.7.9
freezegun==0.3.15
gunicorn==20.0.4
hashids==1.2.0
idna==2.9
inflection==0.4.0
itypes==1.1.0
Jinja2==3.0.0a1
MarkupSafe==2.0.0a1
mccabe==0.6.1
packaging==20.3
pathspec==0.8.0
psycopg2==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
redis==3.5.1
regex==2020.4.4
requests==2.23.0
responses==0.10.14
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sentry-sdk==0.14.3
six==1.14.0
sortedcontainers==2.1.0
sqlparse==0.3.1
tmdbsimple==2.2.0
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
